+++ client/src/reducer/write/writeReducer.type.ts
@@ -1,3 +1,5 @@
+import { FormState } from "types/form.type";
+
 type FormAction =
   | { type: "CHANGE_TITLE"; value: string }
   | { type: "CHANGE_DESCRIPTION"; value: string }
@@ -9,6 +11,7 @@ type FormAction =
   | { type: "DELETE_QUESTION"; questionIndex: number }
   | { type: "COPY_QUESTION"; questionIndex: number }
   | { type: "CHANGE_QUESTION_ESSENTIAL"; questionIndex: number }
-  | { type: "SELECT_FORM_CATEGORY"; value: string };
+  | { type: "SELECT_FORM_CATEGORY"; value: string }
+  | { type: "FETCH_DATA"; init: FormState };
 
 export default FormAction;
